requirements.txt
aiogram==3.15.0
aiohttp==3.10.10
certifi==2024.2.2
fastapi==0.115.6
pydantic==1.10.12
pytest==8.3.3
python-dotenv==1.0.1
python_bcrypt==0.3.2
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
SQLAlchemy==2.0.25
starlette==0.41.3
yarl==1.17.1
